+++ Entities/Set.cs
@@ -1,9 +1,9 @@
 namespace ftDB.Entities
 {
-    public class Set(int weight, int reps, int setNumber, int exerciseInWorkoutId)
+    public class Set(double weight, int reps, int setNumber, int exerciseInWorkoutId)
     {
         public int Id { get; set; }
-        public int Weight { get; set; } = weight;
+        public double Weight { get; set; } = weight;
         public int Reps { get; set; } = reps;
         public int SetNumber { get; set; } = setNumber;
         public int ExerciseInWorkoutId { get; set; } = exerciseInWorkoutId; // FK to ExerciseInWorkout
