+++ Models/Request/PostWorkoutModels/ModelCompletedSet.cs
@@ -2,7 +2,7 @@ using ftDB.BaseLibrary.Models;
 
 namespace ftDB.Models.Request.PostWorkoutModels
 {
-    public class ModelCompletedSet(int weight, int reps, int setNumber, bool isCompleted) : ModelSet(weight, reps, setNumber)
+    public class ModelCompletedSet(double weight, int reps, int setNumber, bool isCompleted) : ModelSet(weight, reps, setNumber)
     {
         public bool IsCompleted { get; set; } = isCompleted;
     }
