+++ Models/Request/PostWorkoutTemplateModels/ModelPostSetTemplate.cs
@@ -2,7 +2,7 @@ using ftDB.BaseLibrary.Models;
 
 namespace ftDB.Models.Request.PostWorkoutTemplateModels
 {
-    public class ModelPostSetTemplate(int weight, int reps, int setNumber) : ModelSet(weight, reps, setNumber)
+    public class ModelPostSetTemplate(double weight, int reps, int setNumber) : ModelSet(weight, reps, setNumber)
     {
 
     }
